readahead: add option to create pack in directory other than root
Add support for creating a readahead pack in a runtime-specified
directory.  Users may want the feature if they their rootfs is
read-only at boot or if they maintain more than one pack file.  The
new pack-file location is specified by a --pack-location command-line
switch.  Default behavior is retained if the switch is absent.